Move ingester error creation in ingester/error.go
