+++ pkg/ingester/ingester_test.go
@@ -6386,7 +6386,7 @@ func TestIngesterUserLimitExceeded(t *testing.T) {
 		httpResp, ok := httpgrpc.HTTPResponseFromError(err)
 		require.True(t, ok, "returned error is not an httpgrpc response")
 		assert.Equal(t, http.StatusBadRequest, int(httpResp.Code))
-		assert.Equal(t, wrapWithUser(makeLimitError(ing.limiter.formatMaxSeriesPerUserError(userID)), userID).Error(), string(httpResp.Body))
+		assert.Equal(t, wrapWithUser(formatMaxSeriesPerUserError(ing.limiter, userID), userID).Error(), string(httpResp.Body))
 
 		// Append two metadata, expect no error since metadata is a best effort approach.
 		_, err = ing.Push(ctx, mimirpb.ToWriteRequest(nil, nil, nil, []*mimirpb.MetricMetadata{metadata1, metadata2}, mimirpb.API))
@@ -6491,7 +6491,7 @@ func TestIngesterMetricLimitExceeded(t *testing.T) {
 		httpResp, ok := httpgrpc.HTTPResponseFromError(err)
 		require.True(t, ok, "returned error is not an httpgrpc response")
 		assert.Equal(t, http.StatusBadRequest, int(httpResp.Code))
-		assert.Equal(t, wrapWithUser(makeMetricLimitError(mimirpb.FromLabelAdaptersToLabels(labels3), ing.limiter.formatMaxSeriesPerMetricError(userID)), userID).Error(), string(httpResp.Body))
+		assert.Equal(t, wrapWithUser(formatMaxSeriesPerMetricError(ing.limiter, mimirpb.FromLabelAdaptersToLabels(labels3), userID), userID).Error(), string(httpResp.Body))
 
 		// Append two metadata for the same metric. Drop the second one, and expect no error since metadata is a best effort approach.
 		_, err = ing.Push(ctx, mimirpb.ToWriteRequest(nil, nil, nil, []*mimirpb.MetricMetadata{metadata1, metadata2}, mimirpb.API))
